Version 2.0
The DQN is working, and if you leave the AI learn for some time, it learns to go to the objective. Lots of work is still needed, and I have to figure out how to speed up the learning, by tweaking the reward, and also make it not collide with walls.